+++ packages/form-render/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gopowerteam/form-render
 
+## 0.0.59
+
+### Patch Changes
+
+- update dependency
+- Updated dependencies
+  - @gopowerteam/modal-render@0.0.38
+
 ## 0.0.58
 
 ### Patch Changes